requirements.txt
dj-database-url==0.4.2
Django==1.11.1
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
gunicorn==19.7.1
mysqlclient==1.3.7
phonenumberslite==8.5.0
pigar==0.6.10
Pillow==3.1.2
pipreqs==0.4.6
psycopg2==2.7.1
pycurl==7.43.0
PyMySQL==0.7.9
whitenoise==3.2.1